Fixes minified JavaScript (JS) by preventing token strings (e.g. `%MY_TOKEN%`) from being evaluated to `true` right away during JS minifcation. Closes #341.
To customize the JS code during runtime we string-replace config tokens of the form `'%MY_TOKEN%'` when serving them to the client via the `JavaScriptServlet` class. Now during the newly introduced JS-minification these tokens were interpreted as `true` which broke this configuration injection. We now declare them as constants which prevents the evaluation.